tests/wpt/web-platform-tests/fullscreen/api/document-onfullscreenchange-manual.html

<!DOCTYPE html>
<title>Document#onfullscreenchange</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<script>
async_test(function(t)
{
    var div = document.querySelector("div");
    assert_equals(document.onfullscreenchange, null, "initial onfullscreenchange");
    document.onfullscreenchange = t.step_func_done();
    trusted_request(t, div);
});
</script>
